### 2025-11-08 #97 @DerKevinRiehl
[¬RE] A Deep Reinforcement Learning Approach for Ramp Metering Based on Traffic Video Data
**Original article:** Liu, Bing, et al. "A deep reinforcement learning approach for ramp metering based on traffic video data." Journal of Advanced Transportation 2021.1 (2021): 6669028. doi: [10.1155/2021/6669028](https://doi.org/10.1155/2021/6669028)

**Scientific domain:** Reinforcement Learning Based Control
**Suggested editor:** [Nicolas P. Rougier](https://github.com/rougier)
