### 2025-12-04 #97 @gdetor
@rougier  Sure, I can handle this.
