### 2025-12-04 #97 @rougier
Thanks !
