### 2026-03-25 #97 @gdetor
@ReScience/reviewers Can anyone handle reviewing this work?
